requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
bilireq==0.1.0a0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
click==8.0.4
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
distlib==0.3.4
fastapi==0.73.0
filelock==3.6.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
haruka-bot==1.3.5
html5lib==1.1
httpcore==0.13.7
httptools==0.3.0
httpx==0.18.2
idna==3.3
importlib-metadata==4.11.2
iso8601==0.1.16
Jinja2==3.0.3
keyring==23.5.0
lockfile==0.12.2
loguru==0.6.0
MarkupSafe==2.1.0
msgpack==1.0.3
multidict==6.0.2
nonebot-adapter-onebot==2.0.0b1
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0b2
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==8.4.0
pip==22.0.4
pkginfo==1.8.2
playwright==1.19.0
poetry==1.1.13
poetry-core==1.0.8
ptyprocess==0.7.0
pyasn1==0.4.8
pydantic==1.9.0
pyee==8.1.0
pygtrie==2.4.2
pylev==1.4.0
pyparsing==3.0.7
pypika-tortoise==0.1.3
python-dotenv==0.17.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
setuptools==57.4.0
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomlkit==0.9.2
tortoise-orm==0.18.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.5
virtualenv==20.13.3
watchgod==0.7
webencodings==0.5.1
websockets==10.1
win32-setctime==1.1.0
yarl==1.7.2
zipp==3.7.0
